backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.3.4
torch>=2.2.0
transformers>=4.44.0
sentence-transformers>=3.0.1
pymilvus>=2.4.4
Pillow>=10.4.0
requests>=2.32.0
tqdm>=4.66.0
numpy>=1.26.0
accelerate>=0.33.0
orjson>=3.10.0
sentencepiece>=0.2.0
tiktoken>=0.8.0
